% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photoswiper.R
\name{psr}
\alias{psr}
\title{Title}
\usage{
psr(data, ...)
}
\arguments{
\item{data}{data.frame with columns
 \itemize{
 \item{src}{: path to image}
 \item{w}{: width}
 \item{h}{: height}
 \item{msrc}{: path to image thumbnail}
 \item{title}{: image caption}
}}

\item{...}{}
}
\description{
Title
}
